requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
joblib==1.3.2
numpy==1.26.2
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2023.3.post1
scikit-learn==0.24.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.3
